tools/jyconvert/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jyconvert"
description = 'Converts file in JSON format to YAML and vice versa'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "sohinimallick16", email = "sohini.mallick@maastrichtuniversity.nl" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["pyyaml"]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/CLARIAH/pure3d/jyconvert#readme"
Issues = "https://github.com/CLARIAH/pure3d/jyconvert/issues"
Source = "https://github.com/CLARIAH/pure3d/jyconvert"

[project.scripts]
jyconvert = "jyconvert.jyconvert:main"

[tool.hatch.version]
path = "jyconvert/jyconvert.py"

[tool.hatch.build.targets.sdist]
[tool.hatch.build.targets.wheel]

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=jyconvert --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "jyconvert/jyconvert.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
